Make SchafKopf ready for KDE 4.9
In libkdegames of KDE 4.9 the following classes we use won't be
included any longer:
KCardDialog, CardDeckInfo, KCardCache, KCardInfo
They have been moved into the repository of the only KDE game
currently using it: lskat.

I copied the source files of these classes from the current trunk
of lskat into our source tree.

Some porting had to be done as the back sides of the cards can't
be selected seperatly any longer.

In the long term we should have a look at libkcardgame which is
currently part of kpat but will go into libkdegames.